# gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VetManager Pro API
  description: |
    API REST para el sistema veterinario SaaS multinacional VetManager Pro.

    **Autenticación:** Bearer JWT (access token de 15min + refresh token de 7 días)

    **Multi-tenant:** Todos los endpoints filtran datos por `branchId` del token JWT.

    **RBAC:** Cada endpoint requiere un permiso específico según el rol del usuario.
  version: "1.0.0"
  contact:
    name: VetManager Pro
    email: api@vetmanagerpro.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.vetmanagerpro.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Pagination:
      type: object
      properties:
        total:      { type: integer }
        page:       { type: integer }
        limit:      { type: integer }
        totalPages: { type: integer }
...
    Client:
      type: object
      properties:
        id:             { type: integer }
        firstName:      { type: string }
        lastName:       { type: string }
        email:          { type: string, format: email }
        phone:          { type: string }
        documentType:   { type: string, enum: [dni, passport, rut, cedula, other] }
        documentNumber: { type: string }
        address:        { type: string }
        city:           { type: string }
        outstandingBalance: { type: number }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /clients:
    get:
      tags: [Clients]
      summary: Listar clientes de la sucursal
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Buscar por nombre, email, teléfono o documento
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Lista paginada de clientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Client' } }
                  meta: { $ref: '#/components/schemas/Pagination' }
    post:
      tags: [Clients]
      summary: Crear nuevo cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, phone]
              properties:
                firstName:      { type: string }
                lastName:       { type: string }
                email:          { type: string, format: email }
                phone:          { type: string }
                documentType:   { type: string, enum: [dni,passport,rut,cedula,other] }
                documentNumber: { type: string }
                address:        { type: string }
                city:           { type: string }
      responses:
        '201': { description: Cliente creado }
        '409': { description: Email o documento duplicado }
  /clients/{id}:
    get:
      tags: [Clients]
      summary: Obtener cliente con sus mascotas
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Cliente con mascotas y contactos de emergencia }
        '404': { description: Cliente no encontrado }
    put:
      tags: [Clients]
      summary: Actualizar cliente
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Actualizado }
    delete:
      tags: [Clients]
      summary: Desactivar cliente (soft delete)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Desactivado }

  # ── PATIENTS ───────────────────────────────────────────────────────────────
...